banner crud setup

+++ app/Http/Controllers/Admin/BannerController.php
@@ -54,7 +54,10 @@ class BannerController extends Controller
             'image' => 'required|image|max:2048', // Ensure it's an image with a size limit
             'agent_id' => $masterCheck ? 'required|exists:users.id' : 'nullable',
         ]);
-        $this->BannerPermission($request->agent_id);
+        $permissionCheck = $this->BannerPermission($request->agent_id);
+        if(!$permissionCheck){
+            return redirect()->back()->with('error', 'You are not authorized to access this feature.');
+        }
         $filename = $this->handleImageUpload($request->image, "banners");
         Banner::create([
             'image' => $filename,
@@ -72,7 +75,10 @@ class BannerController extends Controller
         if (!$banner) {
             return redirect()->back()->with('error', 'Banner Not Found');
         }
-        $this->BannerPermission($banner->agent_id);
+        $permissionCheck = $this->BannerPermission($banner->agent_id);
+        if(!$permissionCheck){
+            return redirect()->back()->with('error', 'You are not authorized to access this feature.');
+        }
         return view('admin.banners.show', compact('banner'));
     }
 
@@ -84,7 +90,10 @@ class BannerController extends Controller
         if (!$banner) {
             return redirect()->back()->with('error', 'Banner Not Found');
         }
-        $this->BannerPermission($banner->agent_id);
+        $permissionCheck = $this->BannerPermission($banner->agent_id);
+        if(!$permissionCheck){
+            return redirect()->back()->with('error', 'You are not authorized to access this feature.');
+        }
         return view('admin.banners.edit', compact('banner'));
     }
 
@@ -97,7 +106,10 @@ class BannerController extends Controller
         if (!$banner) {
             return redirect()->back()->with('error', 'Banner Not Found');
         }
-        $this->BannerPermission($banner->agent_id);
+        $permissionCheck = $this->BannerPermission($banner->agent_id);
+        if(!$permissionCheck){
+            return redirect()->back()->with('error', 'You are not authorized to access this feature.');
+        }
         $request->validate([
             'image' => 'required|image|max:2048', // Ensure it's an image with a size limit
         ]);
@@ -115,7 +127,10 @@ class BannerController extends Controller
         if (!$banner) {
             return redirect()->back()->with('error', 'Banner Not Found');
         }
-        $this->BannerPermission($banner->agent_id);
+        $permissionCheck = $this->BannerPermission($banner->agent_id);
+        if(!$permissionCheck){
+            return redirect()->back()->with('error', 'You are not authorized to access this feature.');
+        }
         $this->handleImageDelete($banner->image, "banners");
         $banner->delete();
         return redirect()->back()->with('success', 'Banner Deleted.');
